Reboot only on POST request; addresses #1324.
git-svn-id: http://svn.freetz.org/trunk@6919 f5190166-0702-4917-9039-51ec32eddaf5

+++ make/mod/files/root/usr/mww/cgi-bin/exec.d/reboot.sh
@@ -2,4 +2,4 @@ cgi_begin '$(lang de:"Neustart" en:"Reboot") ...'
 echo '<p>$(lang de:"Starte neu" en:"Rebooting") ...</p>'
 echo '<p>$(lang de:"Nach dem Neustart <a href=\"/\" target=\"_top\">hier</a> wieder einloggen." en:"Login <a href=\"/\" target=\"_top\">here</a> after reboot.")</p>'
 cgi_end
-reboot
+[ "$REQUEST_METHOD" = POST ] && reboot
